Add mu4e-fill-send-and-exit
This function calls mu4e-fill-buffer-for-flowed and then then calls
messages-send-and-exit.  We previously discussed removing this wrapper
function and moving mu4e-fill-buffer-for-flowed to one of the hooks
invoked while the message is being sent.  However, I decided against
this change to allow this function to take a prefix argument.

With a prefix arg, this function prompts the user for a line lenth to
use when wrapping the paragraphs.  This could be useful for users who
want to wrap their lines differently depending on email recipient; for
instance, I prefer to send lines wrapped at 72 characters when I know
all recipients are likely to be using an email client that supports
plaintext but prefer to send longer lines when sending emails to users
I know to be using Gmail or other webmail clients will poor support
for format=flowed.

See https://vxlabs.com/2019/08/25/format-flowed-with-long-lines/