Update documentation and menus
Note that the current menu implementation is very verbose.  I believe
that this is the correct way to implement a menu letting the user
select from multiple options, but it feels like there ought to be a
better way.

Assuming there isn't, I wonder if it would be worth cutting
format=flowed from the menu altogether and just relying on customize?