Simplify values for mu4e-compose-format-flowed
Previously, `mu4e-compose-format-flowed` took four non-nil values,
some of which were very long/unwieldy.  With this commit, it now takes
three:
 * 'manual
 * 'auto
 * 'auto-long-lines (like 'auto but adds a space to one-line paragraphs)

This drops the ability to refill the buffer based on the `fill-column`
value; now we always use `fill-flowed-encode-column`.  However, this
is not much of a loss -- it is easy to set `fill-flowed-encode-column`
to the value of `fill-column` (and we do so by default).