Fix issue with final paragraph without newline
Previously, a bug existed where the final paragraph would only be
filled if the user put a newline after it.  This bug persisted even
though the code already added a newline before filling the paragraph.

The bug was caused by code called in `message-send-and-exit`, which
messed up the formatting.  This commit fixes the bug by adding our
final newline (if it's needed) before running `message-send-and-exit`
and adding an explanatory comment.