Adding support to query the headers of a message
This commit adds support to search for headers in a message, specified
by its docid. This is necessary, if you want to add support for
protocols, that rely on adding header fields to mails (for example autocrypt).

The mu server is extended by the command "(header :docid <docid>
:header <header>)", and returns a message of the format "(:header
<header-fild> :docid <docid>  :value <header-field-value>)".

mu4e is extended by the function (defun mu4e~proc-header (docid
header) ...), which scans the message with the DOCID for a
HEADER-field.
The result of that function is delivered to the function registered as
"mu4e-search-header-field-func".