Simplify the code for using dired to attach messages
I originally sent some code to the mailing list in 2012 to use dired
for marking attachments.

https://groups.google.com/g/mu-discuss/c/OPwdNZbB5GE/m/hnjNlNoLIu8J

As the function in gnus.el has since been fixed, the version of
gnus-dired-mail-buffers now works fine with mu4e (and has done for
many years).  I suggest deleting the old code to make the
documentation simpler assuming that people are no longer on Emacs from
9 years ago.