mu4e: attempt to surface hung updates in the UI
This is an attempt to give a passive indication of hung email
collections by exposing a status on the main page. The typical failure
case for me is an expired ssh key session leading to background update
hanging because non-interactive sessions shouldn't interrupt the user
for key information. I also occasionally see mbsync hanging during
collection and just never exiting. At least this way I could see an
update started many hours ago and never completed.

However it's apparent that there are other notification and logging
mechanisms which such a thing should maybe use but none seemed to be
good fits. I'm open to a cleaner way to implement this even if it's
just in user hooks.