Add a new maildir section in the homepage.
Since not all the maildirs have a shortcut defined, and the user might want to
see more than the maildirs with a shortcut in their home, we add a new maildir
section in the homepage to see all the defined maildirs to show. The list of
maildirs to be shown are defined in the new custm mu4e-maildirs variable.

Signed-off-by: Ritho <palvarez@ritho.net>