query: include Message-ID header query for related messages
When mu-find is called with a query that includes few emails in a thread, and if
--include-related flag is passed, mu does not find the first email in
thread. This is because in some cases, the first email in the thread does not
includes its own Message-ID header value in the References header. Subsequence
emails include the parent's Message-ID in their own References header, and
that's why mu finds them.

This change modifies the executed query. When searching for related messages, we
include Message-ID search as well. This does pull the first email in the thread.