mu4e: Add mu4e-header-field-face, avoid mu-header-key-face cycle
By default, mu4e-header-key-face inherits from message-header-name; this
leads to warnings when we later remap the latter to the former in
mu4e-compose.el.

We don't want to change the definition of mu4e-header-key-face (it's
used widely in mu4e); so we add a new face mu4e-header-field-face
for *just* mu4e-compose and which does not inherit.

Fixes #2321.