Fix empty 'To:' header on reply to icalendar invitation
On reply to an ical invitation the To: header stayed empty, this fixes that.

The 'Reply-To' data is set if it is not there and if the emaill was different from the Organizer so
that the msg can be further treated by the draft construct functions.

This issue is briefly mentioned in issue #2305