Allow on-the-fly updates to `mu4e-user-mailing-lists' (Bug#2483)
DRAFT.

* mu4e/mu4e-lists.el (mu4e-lists--populate-hash): New helper
function to populate a hash table from an alist.
mu4e-mailing-lists: convert to a `defconst' because it shouldn not
change within an Emacs session.
mu4e-lists--default-hash: new internal constant to hold the hash
table compiled from `mu4e-mailing-lists'.
mu4e-lists--hash: alias from `mu4e--lists-hash', and move it
before the `defcustom'.  Note the TODO.
mu4e-user-mailing-lists: add a :set action to update
`mu4e-lists--hash' when it is modified updated.
(mu4e-get-mailing-list-shortname): remove the now-unnecessary
portion of generating `mu4e-lists--hash' when it is nil.