mu4e-util: allow ESC to exit a choice selection
I occasionally find myself pressing escape to exit a menu choice in mu4e
and it's a bit unfortunate that nothing happens. As best I could tell,
since this is a C call, none of the keymaps are checked for
`(keyboard-quit)' so we need to explicitly check for ESC here

This was previously fixed in 02b5bc805c9b5b7876d7302e172fd1ff9b1ba4aa
but was inadvertently removed in
0354fa4fac08a2799ee50396a035863384303a3c, so more detail was added to
the comment.