Avoid claiming copyright for future years
When compiling today's version in 2039,
the man-pages would state

Copyright © 2022-2039 Dirk-Jan C. Binnema. License GPLv3+: GNU GPL version 3

which cannot be correct, since nobody did any copyrightable work in 2039 in that version.

Additionally, a copyright from 2023 will usually be valid for 70 years
after death of the author, so even if you forget to update it later,
there is no harm.

This patch was done while working on reproducible builds for openSUSE.