man: change quoting style
The man-page sources use single quotes to quote text.  However, this can be
problematic in man-pages because if a single quote appears at the beginning of a
line the following word is interpreted by troff as a macro.  For example, this
paragraph in mu-easy.7:

    What if we want to see some of the body of the message? You can get a 'summary'
    of the first lines of the message using the \fI\-\-summary\-len\fP option, which will
   'summarize' the first \fIn\fP lines of the message:

elicits this warning:

    $ man --warnings obj-x86_64-linux-gnu/man/mu-easy.7 >/dev/null
    troff:<standard input>:166: warning: macro 'summarize'' not defined

and gets truncated:

    What  if  we want to see some of the body of the message? You can get a
    'summary' of the first lines of the message using the --summary-len op‐
    tion, which will

One could adjust the line-wrapping to move the quoted text away from the
beginning of the line, but that is fragile.  Another possibility would be to use
the troff escape-sequences for open and close quotes (`\(oq` and `\(cq`
respectively), but ox-man is being used precisely to avoid having to handle
troff directly.  Instead use back-ticks for left quotes.  Thus:

    What if we want to see some of the body of the message? You can get a `summary'
    of the first lines of the message using the \fI\-\-summary\-len\fP option, which will
   `summarize' the first \fIn\fP lines of the message:

which is rendered correctly:

    What  if  we want to see some of the body of the message? You can get a
    `summary' of the first lines of the message using the --summary-len op-
    tion, which will `summarize' the first n lines of the message:

Signed-off-by: Jeremy Sowden <azazel@debian.org>