man: fix some single-quoted strings
Commit 721aadc140f2 ("man: change quoting style") replaced single-quoted strings
of the form:

    'lorem ipsum'

with:

    `lorem ipsum'

to prevent the possibility of single quotes appearing at the beginning of lines,
because these will be misinterpreted by troff.  However, changes to the man-page
sources in 1.12.9 reintroduced some of these, including one which does appear at
the beginning of a line in mu-query(7):

    '"', '*', '(' and ')'. The details are shell-specific. In case of doubt, the

which results in the following warning:

    troff:<standard input>:25: warning: macro '"',' not defined

and the omission of that line from the rendered paragraph:

    NOTE:  if  you  use  queries on the command-line (say, for mu find), you
    need to quote any characters that would otherwise be interpreted by  the
    shell, such as *--analyze option can be useful.

Use back-ticks for the opening quotes.

Fixes: 5e2b7d52b24d ("mu-query.7: update documentation")
Signed-off-by: Jeremy Sowden <jeremy@azazel.net>